refactor map
