+++ package.json
@@ -21,6 +21,7 @@
     "react-dom": "^16.13.1",
     "react_ujs": "^2.6.1",
     "turbolinks": "^5.2.0",
+    "use-onclickoutside": "^0.3.1",
     "vizzuality-components": "^3.0.1",
     "webpack": "^4.42.0"
   },
